+++ src/connections/api/controllers/state/get/types.d.ts
@@ -1,0 +1,4 @@
+import type { IQuery } from '../../../types';
+import type express from 'express';
+
+export type IGetHealthReq = express.Request<unknown, unknown, undefined, IQuery>;