+++ src/enums/simulation.ts
@@ -1,0 +1,9 @@
+export enum ESimulationStatus {
+  LIVE = 'LIVE',
+  REMOVED = 'REMOVED',
+}
+
+export enum ESimulationCompetitors {
+  HOME = 'HOME',
+  AWAY = 'AWAY',
+}