+++ src/services/state/subModules/get/types.d.ts
@@ -1,0 +1,3 @@
+import type { ISimulationState } from '../../types';
+
+export type IGetSimulationState = Record<string, ISimulationState>;