+++ test/utils/connections.ts
@@ -1,18 +1,20 @@
-import State from '../../src/tools/state';
 import Router from '../../src/connections/api/index';
+import ClientLogic from '../../src/services/state';
 import Bootstrap from '../../src/tools/bootstrap';
+import State from '../../src/tools/state';
 
 export default class Utils {
   constructor() {
-    State.controllers = Bootstrap.getInstance()
-    State.router = new Router()
+    State.controllers = Bootstrap.getInstance();
+    State.router = new Router();
+    State.clientLogic = ClientLogic.getInstance();
   }
 
-  async connect(): Promise<void> {
-    State.router.init()
+  connect(): void {
+    State.router.init();
   }
 
-  async close(): Promise<void> {
+  close(): void {
     State.router.close();
   }
 }