+++ test/utils/setup.ts
@@ -1,24 +1,21 @@
 import { afterAll, beforeAll, vi } from 'vitest';
-import FakeConfigLoader from './fakes/tools/configLoader'
+import FakeConfigLoader from './fakes/tools/configLoader';
 
 vi.mock('../../src/tools/configLoader/index', () => {
-    return {
-      default: FakeConfigLoader
-    };
-  });
-
-import Connections from './connections'
-
-
-const connections = new Connections()
-
-beforeAll(async () => {
-  connections.connect()
-
-
-})
-
-afterAll(() => {
-  connections.close()
+  return {
+    default: FakeConfigLoader,
+  };
 });
 
+// eslint-disable-next-line import/order
+import Connections from './connections';
+
+const connections = new Connections();
+
+beforeAll(() => {
+  connections.connect();
+});
+
+afterAll(() => {
+  connections.close();
+});