+++ changelog.md
@@ -5,6 +5,15 @@ All changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-05-19
+
+### Added
+- Client/state router
+- Global state to hold simulations
+- Logic used to handle api requests
+- Repositories for data
+- Fake classes for better tests
+
 ## [0.1.1] - 2025-05-18
 
 ### Added